pyproject.toml
[project]
name = "review-app"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.34",
  "langfuse>=3.0",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["review_app"]
